[JB][SUPP-689] removing test-data-cor, move test data into test directory
