+++ test/controllers/UpdateEmailVerificationResultControllerSpec.scala
@@ -19,8 +19,8 @@ package controllers
 import essttp.emailverification.EmailVerificationResult
 import essttp.journey.JourneyConnector
 import essttp.rootmodel.IsEmailAddressRequired
-import essttp.testdata.TdAll
 import testsupport.ItSpec
+import testsupport.testdata.TdAll
 
 class UpdateEmailVerificationResultControllerSpec extends ItSpec {
